+++ packages/in-memory/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Add a `reset()` function to in-memory history for resetting the locations. With no arguments, this sets the location to root (`/`) and index to `0`.
+
 ## 1.0.0-beta.5
 
 * Unify `navigate()`, `push()`, and `replace()` under `navigate()`. The optional second argument controls which type of navigation: `ANCHOR` (default), `PUSH`, or `REPLACE`.